refactor: organize core components and services
This commit refactors the project structure by moving shared components, services, and utilities into a new `core` directory. This improves modularity and makes the codebase easier to navigate and maintain.

Key changes:
- Moved gallery_saver, toast, db_helper, log_helper, app_bar, and primary_button into the lib/core directory.
- Updated all import paths to reflect the new file locations.
- Added a core.dart barrel file to simplify imports from the core module.
- Added back_arrow.svg icon.
