+++ lib/core/widgets/app_bar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../core/constants/app_strings.dart';
+import '../constants/app_strings.dart';
 
 enum AppBarType { selection, trash, timeline, detail }
